fix liq buffer to be additive instead of multiplicative

+++ programs/clearing_house/src/state/state.rs
@@ -49,7 +49,7 @@ pub struct State {
     pub max_perp_auction_duration: u8,
     pub min_spot_auction_duration: u8,
     pub max_spot_auction_duration: u8,
-    pub liquidation_margin_buffer_ratio: u8,
+    pub liquidation_margin_buffer_ratio: u32,
     pub signer: Pubkey,
     pub signer_nonce: u8,
